Adds incremental progress on share-file functionality for rooms.
It's been rather frustrating going so far, mainly because http-kit's
multipart support is not where I expect it to be. From what I can tell,
there's limited support for file uploads via the :multipart key but not
much else which makes it difficult to implement HipChat's strict upload
parameters.

I'll either switch to clj-http, which appears to have better multipart
support, or, attempt to submit a PR for http-kit.